activities:data: Refactored directory structure.
