+++ src/main/java/com/owncloud/android/ui/activities/ActivitiesActivity.java
@@ -29,8 +29,8 @@ import com.owncloud.android.lib.resources.files.FileUtils;
 import com.owncloud.android.lib.resources.files.ReadRemoteFileOperation;
 import com.owncloud.android.lib.resources.files.RemoteFile;
 import com.owncloud.android.operations.RefreshFolderOperation;
-import com.owncloud.android.ui.activities.data.ActivitiesServiceApiImpl;
-import com.owncloud.android.ui.activities.data.ActivityRepositories;
+import com.owncloud.android.ui.activities.data.activities.ActivitiesServiceApiImpl;
+import com.owncloud.android.ui.activities.data.activities.ActivityRepositories;
 import com.owncloud.android.ui.activity.FileActivity;
 import com.owncloud.android.ui.activity.FileDisplayActivity;
 import com.owncloud.android.ui.adapter.ActivityListAdapter;
